backend/api_requirements.txt
fastapi[standard]
pillow
requests
qrcode
pydantic_settings
sqlmodel
psycopg[binary,pool]